requirements.txt
numpy
pandas
streamlit
plotly.express
plotly.graph_objects
